ups/requirements.txt
flask>=2.3.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
